% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCMCEnsembleSampler-package.R
\docType{package}
\name{mcmcensemble-package}
\alias{mcmcensemble}
\alias{mcmcensemble-package}
\title{MCMCEnsembleSampler}
\description{
This package implements a particle Monte Carlo Markov Chain sampler
with two different ways of creating proposals.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Bisaloo/mcmcensemble}
  \item \url{https://bisaloo.github.io/mcmcensemble/}
  \item Report bugs at \url{https://github.com/Bisaloo/mcmcensemble/issues}
}

}
\author{
\strong{Maintainer}: Hugo Gruson \email{hugo.gruson+R@normalesup.org} (\href{https://orcid.org/0000-0002-4094-1476}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Sanda Dejanic [copyright holder]
  \item Andreas Scheidegger (\href{https://orcid.org/0000-0003-2575-2172}{ORCID}) [copyright holder]
}

}
\keyword{internal}
