% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCMCEnsembleSampler-package.R
\docType{package}
\name{MCMCEnsembleSampler-package}
\alias{MCMCEnsembleSampler}
\alias{MCMCEnsembleSampler-package}
\title{MCMCEnsembleSampler}
\description{
This package implements a particle Monte Carlo Markov Chain sampler
with two different ways of creating proposals.
}
\author{
\strong{Maintainer}: Hugo Gruson \email{hugo.gruson+R@normalesup.org} (\href{https://orcid.org/0000-0002-4094-1476}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Sanda Dejanic [copyright holder]
  \item Andreas Scheidegger (\href{https://orcid.org/0000-0003-2575-2172}{ORCID}) [copyright holder]
}

}
\keyword{internal}
